Refactor inbox summary into separate module
